requirements.txt
openai-whisper
numpy
pandas
requests
python-dotenv
pyaudio
soundfile
pyttsx3
openai